Make obs fields floats. refs #2

+++ obs/models.py
@@ -9,12 +9,12 @@ class Observation(EpisodeSubrecord):
     _batch_template = True
     _sort           = 'date'
 
-    bp_systolic  = models.IntegerField(blank=True, null=True)
-    bp_diastolic = models.IntegerField(blank=True, null=True)
-    pulse        = models.IntegerField(blank=True, null=True)
-    resp_rate    = models.IntegerField(blank=True, null=True)
-    sp02         = models.IntegerField(blank=True, null=True)
-    temperature  = models.IntegerField(blank=True, null=True)
-    height       = models.IntegerField(blank=True, null=True)
-    weight       = models.IntegerField(blank=True, null=True)
+    bp_systolic  = models.FloatField(blank=True, null=True)
+    bp_diastolic = models.FloatField(blank=True, null=True)
+    pulse        = models.FloatField(blank=True, null=True)
+    resp_rate    = models.FloatField(blank=True, null=True)
+    sp02         = models.FloatField(blank=True, null=True)
+    temperature  = models.FloatField(blank=True, null=True)
+    height       = models.FloatField(blank=True, null=True)
+    weight       = models.FloatField(blank=True, null=True)
     date         = models.DateField(blank=True, null=True)
